Add benchmarks
This base58 library uses big.Int instead of []byte, and will never be
blazingly fast for larger inputs. It is more meant for encoding
numeric ids, such as what you might have in your database.

For generic encoding routines, more in the style of base64 and hex,
and a lot more speed, see https://github.com/cmars/basen

This library also is simple enough, and contains multiple commands, to
serve as a good example of repository organization; if this weren't
true, I would probably merge it with cmars/basen.